pyproject.toml
[project]
name = "infosec-labs"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=11.2.1",
    "typer>=0.15.4",
]

[dependency-groups]
dev = [
    "black>=25.1.0",
    "mypy>=1.15.0",
    "python-lsp-server[all]>=1.12.2",
    "ruff>=0.11.10",
]
